math/ld: Fix implicit casts
Add a pile of explicit casts and increase precision of a couple of
constants.

+++ newlib/libm/complex/cacosl.c
@@ -41,7 +41,7 @@ cacosl(long double complex z)
 	long double complex w;
 
 	w = casinl(z);
-	w = (M_PI_2L - creall(w)) - cimagl(w) * (double complex) I;
+	w = (M_PI_2L - creall(w)) - cimagl(w) * (long double complex) I;
 	return w;
 }
