+++ newlib/libm/complex/catanhl.c
@@ -38,7 +38,7 @@ catanhl(long double complex z)
 {
 	long double complex w;
 
-	w = -1.0L * (double complex) I * catanl(z * (double complex) I);
+	w = -1.0L * (long double complex) I * catanl(z * (long double complex) I);
 	return w;
 }
