+++ newlib/libm/complex/ccoshl.c
@@ -42,7 +42,7 @@ ccoshl(long double complex z)
 
 	x = creall(z);
 	y = cimagl(z);
-	w = coshl(x) * cosl(y) + (sinhl(x) * sinl(y)) * (double complex) I;
+	w = coshl(x) * cosl(y) + (sinhl(x) * sinl(y)) * (long double complex) I;
 	return w;
 }
