+++ newlib/libm/complex/cexpl.c
@@ -43,7 +43,7 @@ cexpl(long double complex z)
 	x = creall(z);
 	y = cimagl(z);
 	r = expl(x);
-	w = r * cosl(y) + r * sinl(y) * (double complex) I;
+	w = r * cosl(y) + r * sinl(y) * (long double complex) I;
 	return w;
 }
