+++ newlib/libm/complex/clogl.c
@@ -43,7 +43,7 @@ clogl(long double complex z)
 	rr = cabsl(z);
 	p = logl(rr);
 	rr = atan2l(cimagl(z), creall(z));
-	w = p + rr * (double complex) I;
+	w = p + rr * (long double complex) I;
 	return w;
 }
