+++ newlib/libm/complex/ctanhl.c
@@ -43,7 +43,7 @@ ctanhl(long double complex z)
 	x = creall(z);
 	y = cimagl(z);
 	d = coshl(2.0L * x) + cosl(2.0L * y);
-	w = sinhl(2.0L * x) / d  +  (sinl(2.0L * y) / d) * (double complex) I;
+	w = sinhl(2.0L * x) / d  +  (sinl(2.0L * y) / d) * (long double complex) I;
 
 	return w;
 }
